storage: split user manager
