+++ backend/storage/__init__.py
@@ -4,6 +4,8 @@ Storage module for persistent file-based storage of user preferences.
 
 from .storage_manager import StorageManager
 from .user_manager import UserManager
+from .profile_manager import ProfileManager
+from .research_manager import ResearchManager
 from .zep_manager import ZepManager
 
-__all__ = ['StorageManager', 'UserManager', 'ZepManager'] 
+__all__ = ["StorageManager", "UserManager", "ProfileManager", "ResearchManager", "ZepManager"]
